2.10.0 升级利刃3.7.0 修改密码页面支持C端

+++ uap/src/main/java/com/dili/uap/api/UserApi.java
@@ -1,18 +1,5 @@
 package com.dili.uap.api;
 
-import java.util.List;
-import com.dili.uap.glossary.UserState;
-import com.dili.uap.service.LoginService;
-import org.apache.commons.lang.StringEscapeUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.dili.ss.domain.BaseOutput;
@@ -25,11 +12,21 @@ import com.dili.uap.domain.ScheduleMessage;
 import com.dili.uap.domain.dto.UserDepartmentRole;
 import com.dili.uap.domain.dto.UserDepartmentRoleQuery;
 import com.dili.uap.domain.dto.UserDto;
+import com.dili.uap.glossary.UserState;
 import com.dili.uap.sdk.domain.User;
 import com.dili.uap.sdk.domain.dto.UserQuery;
+import com.dili.uap.service.LoginService;
 import com.dili.uap.service.RoleService;
 import com.dili.uap.service.UserService;
 import com.github.pagehelper.Page;
+import org.apache.commons.lang.StringEscapeUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 
 /**
@@ -106,9 +103,9 @@ public class UserApi {
 		if (users instanceof Page) {
 			Page<User> page = (Page) users;
 			Long total = page.getTotal();
-			return PageOutput.success().setTotal(total.intValue()).setPageNum(page.getPageNum()).setPageSize(page.getPageSize()).setData(users);
+			return PageOutput.success().setTotal(total).setPageNum(page.getPageNum()).setPageSize(page.getPageSize()).setData(users);
 		} else {
-			return PageOutput.success().setTotal(users.size()).setPageNum(user.getPage()).setPageSize(user.getRows()).setData(users);
+			return PageOutput.success().setTotal((long)users.size()).setPageNum(user.getPage()).setPageSize(user.getRows()).setData(users);
 		}
 	}
